+++ packages/editor/src/utils/is.ts
@@ -1,0 +1,5 @@
+export const isContentEditable = (
+  target: EventTarget | Element | null,
+): target is HTMLDivElement => {
+  return target instanceof HTMLDivElement && target.isContentEditable;
+};