+++ example/CHANGELOG.md
@@ -1,5 +1,22 @@
 # example
 
+## 0.5.3
+
+### Patch Changes
+
+- 3c3bf37: perf tuning
+- Rename code block to block code
+- 2a464bc: - Add local storage plugin
+  - Add baisc image block template
+  - Add placeholder button component
+- Updated dependencies [3c3bf37]
+- Updated dependencies [undefined]
+- Updated dependencies [2a464bc]
+  - @hexx/editor@0.6.5
+  - @hexx/renderer@0.6.5
+  - @hexx/theme@0.6.5
+  - @hexx/block-code@0.6.5
+
 ## 0.5.2
 
 ### Patch Changes