+++ packages/block-code/CHANGELOG.md
@@ -1,4 +1,14 @@
-# @hexx/code-block
+# @hexx/block-code
+
+## 0.6.5
+
+### Patch Changes
+
+- Rename code block to block code
+- Updated dependencies [3c3bf37]
+- Updated dependencies [undefined]
+- Updated dependencies [2a464bc]
+  - @hexx/editor@0.6.5
 
 ## 1.0.3